Separate ProtectedRoute and AdminProtectedRoute components into their own files
App should only consist of the App component.
