+++ app/imports/ui/layouts/App.jsx
@@ -1,9 +1,8 @@
 import React from 'react';
-import PropTypes from 'prop-types';
-import { Meteor } from 'meteor/meteor';
 import { useTracker } from 'meteor/react-meteor-data';
 import { Roles } from 'meteor/alanning:roles';
-import { BrowserRouter as Router, Route, Routes, Navigate } from 'react-router-dom';
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
+
 import Footer from '../components/Footer';
 import Landing from '../pages/Landing';
 import ListStuff from '../pages/ListStuff';
@@ -16,7 +15,8 @@ import SignOut from '../pages/SignOut';
 import NavBar from '../components/NavBar';
 import SignIn from '../pages/SignIn';
 import NotAuthorized from '../pages/NotAuthorized';
-import LoadingSpinner from '../components/LoadingSpinner';
+import ProtectedRoute from './ProtectedRoute';
+import AdminProtectedRoute from './AdminProtectedRoute';
 
 /** Top-level layout component for this application. Called in imports/startup/client/startup.jsx. */
 const App = () => {
@@ -47,53 +47,6 @@ const App = () => {
       </div>
     </Router>
   );
-};
-
-/*
- * ProtectedRoute (see React Router v6 sample)
- * Checks for Meteor login before routing to the requested page, otherwise goes to signin page.
- * @param {any} { component: Component, ...rest }
- */
-const ProtectedRoute = ({ children }) => {
-  const isLogged = Meteor.userId() !== null;
-  return isLogged ? children : <Navigate to="/signin" />;
-};
-
-/**
- * AdminProtectedRoute (see React Router v6 sample)
- * Checks for Meteor login and admin role before routing to the requested page, otherwise goes to signin page.
- * @param {any} { component: Component, ...rest }
- */
-const AdminProtectedRoute = ({ ready, children }) => {
-  const isLogged = Meteor.userId() !== null;
-  if (!isLogged) {
-    return <Navigate to="/signin" />;
-  }
-  if (!ready) {
-    return <LoadingSpinner />;
-  }
-  const isAdmin = Roles.userIsInRole(Meteor.userId(), 'admin');
-  return (isLogged && isAdmin) ? children : <Navigate to="/notauthorized" />;
-};
-
-// Require a component and location to be passed to each ProtectedRoute.
-ProtectedRoute.propTypes = {
-  children: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
-};
-
-ProtectedRoute.defaultProps = {
-  children: <Landing />,
-};
-
-// Require a component and location to be passed to each AdminProtectedRoute.
-AdminProtectedRoute.propTypes = {
-  ready: PropTypes.bool,
-  children: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
-};
-
-AdminProtectedRoute.defaultProps = {
-  ready: false,
-  children: <Landing />,
 };
 
 export default App;
