+++ app/imports/ui/layouts/ProtectedRoute.jsx
@@ -1,0 +1,28 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { Meteor } from 'meteor/meteor';
+import { Navigate } from 'react-router';
+
+import Landing from '../pages/Landing';
+
+/**
+ * ProtectedRoute (see React Router v6 sample)
+ * Checks for Meteor login before routing to the requested page, otherwise goes to signin page.
+ *
+ * @type { React.FC<{ children: React.ReactNode }> }
+ */
+const ProtectedRoute = ({ children }) => {
+  const isLogged = Meteor.userId() !== null;
+  return isLogged ? children : <Navigate to="/signin" />;
+};
+
+// Require a component and location to be passed to each ProtectedRoute.
+ProtectedRoute.propTypes = {
+  children: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
+};
+
+ProtectedRoute.defaultProps = {
+  children: <Landing />,
+};
+
+export default ProtectedRoute;
